put chess button class as an entity so it can be accessed throughout the program according to CA
